mountd: add support for case-insensitive file names
Case insensitive filesystems support textually distinct names for the
same directory. i.e. you can access it with a name other than the
canonical name.
For example if you
  mkdir /mnt/export

then add /mnt/EXPORT to /etc/exports, and on a client
  mount server:/mnt/EXPORT /import

then the mount will work, but if the kernel on the server needs to
refresh the export information, it will ask about "/mnt/export", which
is not listed in /etc/exports and so will fail.

To fix this we need mountd to perform case-insensitive name
comparisons, but only when the filesystem would, and in exactly the
same way that the filesystem would.

So, when comparing paths for equality first try some simple heuristics
which will not be fooled by case and then ask the kernel if they are
the same.

By preference we use name_to_handle_at() as it reports the mntid which
can distinguish between bind mounts.  If that is not available, use
lstat() and compare rdev and ino.

Acked-by: J. Bruce Fields <bfields@fieldses.org>
Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Steve Dickson <steved@redhat.com>