blkmapd: fix broken multipath handling
We do want to use the dm-multipath device if it exists, which the
code is generally prepared for, except that this check excludes them
early.  In addition this will also add the passive path to the device
list, which is harmless if an active one exists as that or the multipath
device will be preferred, and at least allows us to work if it doesn't.

Also fix up the check if an path needs to be updated to remove the silly
partition check - pNFS block offset are relative to the device so partion
should never match it instead of the full device.  On the other hand the
simplistic check easily creates false positives, e.g. dm-10 is
considered a partition of dm-1.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Steve Dickson <steved@redhat.com>