mount.nfs: configurable minor version defaults
Update nfsmount.conf to allow minor version specification, and rearrange
the autonegotiation logic to agreed upon best behavior.  Depending upon the
combination of values specified within nfsmount.conf and given to mount.nfs,
the retry behavior of mount.nfs varies according to the general rule of
defaulting to the most specific setting, with some exceptions.  A
general guide to the expected behavior follows:

------------------
| nfsmount.conf  |-----------------------------------
| Defaultvers=   |  arg option  |     attempts:     |
|---------------------------------------------------|
|     4.2        |  not set     | v4.2 v4.1 v4.0 v3 |
|     4.2        |   v4         | v4.2 v4.1 v4.0    |
|     4.1        |  not set     |      v4.1 v4.0 v3 |
|     4.1        |   v4         |      v4.1 v4.0    |
|     4          |  not set     |           v4.0 v3 |
|     4          |   v4         |           v4.0    |
|     3          |  not set     |                v3 |
|    any set     |   v4.2       | v4.2              |
|    any set     |   v4.1       |      v4.1         |
|    any set     |   v4         |           v4.0    |
|    any set     |   v3         |                v3 |
|    not set     |  not set     | v4.2 v4.1 v4.0 v3 |
-----------------------------------------------------

If built with --enable-mountconfig=no, then the behavior is the same as if
nfsmount.conf did not set Defaultvers.

Signed-off-by: Benjamin Coddington <bcodding@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>