cleanup daemonization code
The daemonization init/ready functions have parameters that are never used,
require the caller to keep track of some pipefds that it has no interest in
and which might not be used in some scenarios. Cleanup both functions a bit.

The idea here is also that these two functions might be good points to
insert more systemd init code later (sd_notify()).

Also, statd had a private copy of the daemonization code for unknown
reasons...so make it use the generic version instead.

Signed-off-by: David H?rdeman <david@hardeman.nu>
Signed-off-by: Steve Dickson <steved@redhat.com>