mountstats: Fix exit code handling
When I changed mountstats to use the argparse module, I neglected to
make the subcommand functions return any values even though main() was
looking for them.  Also removed SystemExit from the except clause at
the end of the program since it was causing the program to always exit
with a status of 1.

Signed-off-by: Scott Mayhew <smayhew@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>