close the syslog fd in daemon_init()
Commit 7addf9d (cleanup daemonization code) added the following line to
mydaemon_init():

        dup2(pipefds[1], 3);

If we've already called vsyslog() before the fork(), then chances are fd
3 was being used for the syslog socket.  In that case the next vsyslog()
call will cause the data to appear on the read end of the pipe, causing
the parent to exit with a nonzero status.  If systemd is running, it
will see the parent's nonzero exit status and will terminate the child
as well.

So just call closelog() to close the fd.  The next call to vsyslog()
will open a new one if need be.

Signed-off-by: Scott Mayhew <smayhew@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>