systemd: ensure nfs-config service is re-run as needed.
The nfs-config service translates distro-specific startup
configuration into "environment" variable read and used
by systemd unit files.

Currently it is only run once, so subsequent changes to the
distro-specific files do not take effect when an nfs service is
restarted.

If we change "RemainAfterExit=yes" to "RemainAfterExit=no" then the
service will be restarted before any dependant service is started, so
the environment file will always be up to date.

Reported-and-tested-by: Benjamin Coddington <bcodding@redhat.com>
Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Steve Dickson <steved@redhat.com>