mount: run START_STATD fully as root
If a "user" mount is the first NFSv3 mount, mount.nfs will be running
setuid to root (with non-root as the real-uid) when it executes
START_STATD.

start-statd is a shell script and many shells refuse to run setuid,
dropping privileges immediately.  This results in start-statd running
as an unprivileged user and so statd fails to start.

To fix this, call "setuid(0)" to set real uid to zero.  Also call
"setgid(0)"
for consistency.

The behaviour of a shell can often be affected by the environment,
such as the "shell functions" that bash includes from the environment.
To avoid the user being able to pass such environment to the shell,
explicitly pass an empty environment.  The start-statd script explicitly
sets the PATH which is all it really needs.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Steve Dickson <steved@redhat.com>