gssd: Convert 'rdma' to 'tcp' protocol
The kernel NFS client specifies 'rdma' as the transport protocol for
NFS/RDMA mounts. This is a generic display token, probably not
something the kernel should lie about. The kernel uses it to form
NFSv4 client IDs, for example.

Until gssd can speak RPC-over-RDMA, convert 'rdma' to 'tcp' in
kernel gss upcalls.

This means that the Linux NFS client requires an NFS/TCP service on
an NFS/RDMA server when establishing a GSS context for proto=rdma
mount points. This is probably not an onerous requirement for the
time being.

Eventually user space support for RPC-over-RDMA might be introduced,
and gssd can use that for context establishment.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>