mount: RPC_PROGNOTREGISTERED should not be a permanent error
Commit: bf66c9facb8e ("mounts.nfs: v2 and v3 background mounts should
retry when server is down.")

changed the behaviour of "bg" mounts so that RPC_PROGNOTREGISTERED,
which maps to EOPNOTSUPP, is not a permanent error.
This useful because when an NFS server starts up there is a small window
between the moment that rpcbind (or portmap) starts responding to lookup
requests, and the moment when nfsd registers with rpcbind.  During that window
rpcbind will reply with RPC_PROGNOTREGISTERED, but mount should not give
up.

This same reasoning applies to foreground mounts.  They don't wait for
as long, but could still hit the window and fail prematurely.

So revert the above patch and instead add EOPNOTSUPP to the list of
temporary errors known to nfs_is_permanent_error.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Steve Dickson <steved@redhat.com>