nfs-server-generator: avoid using external services.
nfs-server-generator is run very early when a lot of services are not
yet started, so it mustn't depend on them.  Currently it can try to
use hostname lookup and syslog.  Using hostname lookup can cause
errors and when these are logged via syslog, it can cause the
generator to block indefinitely

Hostname-lookup is not needed, as we don't use the host issue,
and sending message to stderr is sufficient for the generator.

Disabling syslog is easy - call a function that sets a static variable.

Disabling hostname lookup requires adding an "ignore_hosts" flags to
export_read and export_d_read().

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Steve Dickson <steved@redhat.com>